feat: allow authenticate_as_user for user JWT tokens
Some API calls mandate that a user is used for the API authentication,
whilst others allow it optionally. Instead of trying to figure out which
ones allow it and which don't, the user can provide an
`authenticate_as_user` option, which if provided, will use a user
authentication token instead of a server one.